Replace terminal commands with scripts
Replace the one-line terminal commands provided in the
'two-dogs/the-kennel' repository with script alternatives.

As these commands are intended to be used by newcomers to Linux Mint,
requiring copy/paste to a terminal window may introduce additional
errors and frustration for the user.

Removed *md files that have been replaced with scripts.